requirements.txt
pandas
numpy
pyyaml
loguru
requests
tabulate
scikit-learn
lightgbm
joblib
